Basic KNN Recommender
This recommender uses the 100K MovieLens dataset - ratings.csv and movies_sm.csv are the files to use with notebook. I'll upload them. The recommender is still very basic. The code at the end just randomly selects a movie and makes a top 5 recommendation for that movie. Will extend the functionality later.